requirements.txt
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
python-dotenv>=1.0.0
pytest>=7.3.1
pytest-django>=4.5.2
PyYAML>=6.0.0
inflection>=0.5.1
uritemplate>=4.1.1
paho-mqtt
channels>=4.0.0
daphne>=4.0.0
celery>=5.4.0
Pillow>=10.0.0
django-filter>=23.2
django-cors-headers>=4.2.0
psycopg2-binary>=2.9.6
gunicorn>=21.2.0
